requirements.txt
bcrypt==5.0.0
blinker==1.9.0
click==8.3.1
exceptiongroup==1.3.1
Flask==3.1.3
flask-cors==6.0.2
greenlet==3.3.1
gunicorn==25.1.0
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
pytest==9.0.2
python-dotenv==1.2.1
SQLAlchemy==2.0.46
tomli==2.4.0
typing_extensions==4.15.0
Werkzeug==3.1.6
